fix(v2.1): reject RV64 load/store access to deferral address space (#2921)
## Summary

- reject RV64 loads unless they read from user memory (`e = 2`)
- reject RV64 stores unless they write to user memory (`e = 2`) or
public values (`e = 3`)
- remove the obsolete AS 4 store pseudo-instruction path so the deferral
address space is only maintained by the deferral extension
- update RVR lifting so public-values stores are handled by the IO
extension, including byte/half/word/dword REVEAL widths
- improve invalid address-space error messages for load/store preflight
paths
- store load/store `prev_data` as bytes now that this path no longer
supports field-cell deferral writes
- update ISA/deferral docs and add regression coverage for invalid
deferral-address-space load/store usage

resolves int-7698

+++ extensions/riscv/circuit/cuda/src/loadstore.cu
@@ -30,8 +30,7 @@ template <size_t NUM_CELLS> struct LoadStoreCoreRecord {
     uint8_t local_opcode;
     uint8_t shift_amount;
     uint8_t read_data[NUM_CELLS];
-    // Note: `prev_data` can be a field, so we need to use u32
-    uint32_t prev_data[NUM_CELLS];
+    uint8_t prev_data[NUM_CELLS];
 };
 
 enum Rv64LoadStoreOpcode {
@@ -84,10 +83,10 @@ __device__ constexpr uint32_t LOADSTORE_WIDTH[] = {
 
 template <size_t NUM_CELLS>
 __device__ __forceinline__ void run_write_data(
-    uint32_t (&write_data)[NUM_CELLS],
+    uint8_t (&write_data)[NUM_CELLS],
     Rv64LoadStoreOpcode opcode,
     const uint8_t (&read_data)[NUM_CELLS],
-    const uint32_t (&prev_data)[NUM_CELLS],
+    const uint8_t (&prev_data)[NUM_CELLS],
     uint8_t shift
 ) {
     bool is_store = opcode >= STORED;
@@ -123,7 +122,7 @@ template <size_t NUM_CELLS> struct LoadStoreCore {
             true
         );
         uint8_t shift = record.shift_amount;
-        uint32_t write_data[NUM_CELLS] = {0};
+        uint8_t write_data[NUM_CELLS] = {0};
 
         COL_WRITE_VALUE(row, Cols, is_valid, 1);
         COL_WRITE_VALUE(
